refactor(tag): move tag into it's own top-level module
