+++ src/mvba/mvba/message.rs
@@ -2,7 +2,7 @@ use blsttc::{Signature, SignatureShare};
 use serde::{Deserialize, Serialize};
 
 use super::NodeId;
-use crate::mvba::{hash::Hash32, vcbc::message::Tag};
+use crate::mvba::{hash::Hash32, tag::Tag};
 
 /// VoteAction definition.
 /// This is same as `v-vote` message in spec: (ID, v-vote, a, uj, ρj)
